requirements.txt
Pillow
PyYAML
addict
future
insightface>=0.7.3
lmdb
lpips
numpy
onnxruntime
opencv-python
pillow>=9.5.0
protobuf>=4.23.2
scikit-image
scipy
tb-nightly
torch
torchvision
tqdm
yapf
